clean project

+++ Cs4rsaDatabaseService/Models/Curriculum.cs
@@ -4,7 +4,7 @@ namespace Cs4rsaDatabaseService.Models
 {
     public class Curriculum
     {
-        public int CurriculumId { get; set; }
+        public short CurriculumId { get; set; }
         public string Name { get; set; }
         public List<Student> Students { get; set; }
     }
